+++ examples/github-complete/outputs.tf
@@ -4,9 +4,9 @@ output "atlantis_url" {
   value       = module.atlantis.atlantis_url
 }
 
-output "atlantis_allowed_repo_names" {
+output "atlantis_repo_allowlist" {
   description = "Git repositories where webhook should be created"
-  value       = module.atlantis.atlantis_allowed_repo_names
+  value       = module.atlantis.atlantis_repo_allowlist
 }
 
 output "task_role_arn" {
@@ -22,10 +22,11 @@ output "ecs_task_definition" {
 # Webhooks
 output "github_webhook_urls" {
   description = "Github webhook URL"
-  value       = module.github_repository_webhook.this_repository_webhook_urls
+  value       = module.github_repository_webhook.repository_webhook_urls
 }
 
 output "github_webhook_secret" {
   description = "Github webhook secret"
-  value       = module.github_repository_webhook.this_repository_webhook_secret
+  value       = module.github_repository_webhook.repository_webhook_secret
+  sensitive   = true
 }