+++ examples/github-repository-webhook/outputs.tf
@@ -1,9 +1,10 @@
 output "github_webhook_urls" {
   description = "Github webhook URL"
-  value       = module.github_repository_webhook.this_repository_webhook_urls
+  value       = module.github_repository_webhook.repository_webhook_urls
 }
 
 output "github_webhook_secret" {
   description = "Github webhook secret"
-  value       = module.github_repository_webhook.this_repository_webhook_secret
+  value       = module.github_repository_webhook.repository_webhook_secret
+  sensitive   = true
 }