+++ main.tf
@@ -5,11 +5,11 @@ locals {
   public_subnet_ids  = coalescelist(module.vpc.public_subnets, var.public_subnet_ids, [""])
 
   # Atlantis
-  atlantis_image = var.atlantis_image == "" ? "runatlantis/atlantis:${var.atlantis_version}" : var.atlantis_image
+  atlantis_image = var.atlantis_image == "" ? "ghcr.io/runatlantis/atlantis:${var.atlantis_version}" : var.atlantis_image
   atlantis_url = "https://${coalesce(
     var.atlantis_fqdn,
     element(concat(aws_route53_record.atlantis.*.fqdn, [""]), 0),
-    module.alb.this_lb_dns_name,
+    module.alb.lb_dns_name,
     "_"
   )}"
   atlantis_url_events = "${local.atlantis_url}/events"
@@ -25,7 +25,10 @@ locals {
   secret_webhook_key = local.has_secrets || var.atlantis_github_webhook_secret != "" ? var.atlantis_gitlab_user_token != "" ? "ATLANTIS_GITLAB_WEBHOOK_SECRET" : var.atlantis_github_user_token != "" || var.atlantis_github_webhook_secret != "" ? "ATLANTIS_GH_WEBHOOK_SECRET" : "ATLANTIS_BITBUCKET_WEBHOOK_SECRET" : ""
 
   # determine if the alb has authentication enabled, otherwise forward the traffic unauthenticated
-  alb_authenication_method = length(keys(var.alb_authenticate_oidc)) > 0 ? "authenticate-oidc" : length(keys(var.alb_authenticate_cognito)) > 0 ? "authenticate-cognito" : "forward"
+  alb_authentication_method = length(keys(var.alb_authenticate_oidc)) > 0 ? "authenticate-oidc" : length(keys(var.alb_authenticate_cognito)) > 0 ? "authenticate-cognito" : "forward"
+
+  # ECS - existing or new?
+  ecs_cluster_id = var.create_ecs_cluster ? module.ecs.ecs_cluster_id : var.ecs_cluster_id
 
   # Container definitions
   container_definitions = var.custom_container_definitions == "" ? var.atlantis_bitbucket_user_token != "" ? jsonencode(concat([module.container_definition_bitbucket.json_map_object], var.extra_container_definitions)) : jsonencode(concat([module.container_definition_github_gitlab.json_map_object], var.extra_container_definitions)) : var.custom_container_definitions
@@ -109,20 +112,41 @@ locals {
     },
     var.tags,
   )
+
+  policies_arn = var.policies_arn != null ? var.policies_arn : ["arn:${data.aws_partition.current.partition}:iam::aws:policy/service-role/AmazonECSTaskExecutionRolePolicy"]
+
+  # Chunk these into groups of 5, the limit for IPs in an AWS lb listener
+  whitelist_unauthenticated_cidr_block_chunks = chunklist(
+    sort(compact(concat(var.allow_github_webhooks ? var.github_webhooks_cidr_blocks : [], var.whitelist_unauthenticated_cidr_blocks))),
+    5
+  )
+
+  # break up user to uid and gid -- set both to 0 if null
+  uid = var.user == null ? 0 : split(":", var.user)[0]
+  gid = var.user == null ? 0 : split(":", var.user)[1]
+
+  # default mount points for efs if ephemeral storage is not enabled and mount points aren't specified
+  mount_points = var.enable_ephemeral_storage || length(var.mount_points) > 0 ? var.mount_points : [{
+    containerPath = "/home/atlantis"
+    sourceVolume  = "efs-storage"
+    readOnly      = "false"
+  }]
 }
+
+data "aws_partition" "current" {}
 
 data "aws_region" "current" {}
 
 data "aws_route53_zone" "this" {
-  count = var.create_route53_record ? 1 : 0
+  count = var.create_route53_record || var.create_route53_aaaa_record ? 1 : 0
 
   name         = var.route53_zone_name
-  private_zone = false
+  private_zone = var.route53_private_zone
 }
 
-###################
+################################################################################
 # Secret for webhook
-###################
+################################################################################
 resource "random_id" "webhook" {
   count = var.atlantis_github_webhook_secret != "" ? 0 : 1
 
@@ -169,12 +193,13 @@ resource "aws_ssm_parameter" "atlantis_bitbucket_user_token" {
   tags = local.tags
 }
 
-###################
+################################################################################
 # VPC
-###################
+################################################################################
+
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "v2.64.0"
+  version = "v3.6.0"
 
   create_vpc = var.vpc_id == ""
 
@@ -185,31 +210,38 @@ module "vpc" {
   private_subnets = var.private_subnets
   public_subnets  = var.public_subnets
 
-  enable_nat_gateway = true
-  single_nat_gateway = true
+  enable_nat_gateway   = true
+  single_nat_gateway   = true
+  enable_dns_hostnames = !var.enable_ephemeral_storage
+
+  manage_default_security_group  = var.manage_default_security_group
+  default_security_group_ingress = var.default_security_group_ingress
+  default_security_group_egress  = var.default_security_group_egress
 
   tags = local.tags
 }
 
-###################
+################################################################################
 # ALB
-###################
+################################################################################
 module "alb" {
   source  = "terraform-aws-modules/alb/aws"
-  version = "v5.10.0"
+  version = "v6.5.0"
 
   name     = var.name
   internal = var.internal
 
   vpc_id          = local.vpc_id
   subnets         = local.public_subnet_ids
-  security_groups = flatten([module.alb_https_sg.this_security_group_id, module.alb_http_sg.this_security_group_id, var.security_group_ids])
+  security_groups = flatten([module.alb_https_sg.security_group_id, module.alb_http_sg.security_group_id, var.security_group_ids])
 
   access_logs = {
     enabled = var.alb_logging_enabled
     bucket  = var.alb_log_bucket_name
     prefix  = var.alb_log_location_prefix
   }
+
+  ip_address_type = var.alb_ip_address_type
 
   enable_deletion_protection = var.alb_enable_deletion_protection
 
@@ -221,8 +253,8 @@ module "alb" {
       target_group_index   = 0
       port                 = 443
       protocol             = "HTTPS"
-      certificate_arn      = var.certificate_arn == "" ? module.acm.this_acm_certificate_arn : var.certificate_arn
-      action_type          = local.alb_authenication_method
+      certificate_arn      = var.certificate_arn == "" ? module.acm.acm_certificate_arn : var.certificate_arn
+      action_type          = local.alb_authentication_method
       authenticate_oidc    = var.alb_authenticate_oidc
       authenticate_cognito = var.alb_authenticate_cognito
     },
@@ -248,6 +280,9 @@ module "alb" {
       backend_port         = var.atlantis_port
       target_type          = "ip"
       deregistration_delay = 10
+      health_check = {
+        path = "/healthz"
+      }
     },
   ]
 
@@ -256,10 +291,10 @@ module "alb" {
 
 # Forward action for certain CIDR blocks to bypass authentication (eg. GitHub webhooks)
 resource "aws_lb_listener_rule" "unauthenticated_access_for_cidr_blocks" {
-  count = var.allow_unauthenticated_access ? 1 : 0
+  count = var.allow_unauthenticated_access ? length(local.whitelist_unauthenticated_cidr_block_chunks) : 0
 
   listener_arn = module.alb.https_listener_arns[0]
-  priority     = var.allow_unauthenticated_access_priority
+  priority     = var.allow_unauthenticated_access_priority + count.index
 
   action {
     type             = "forward"
@@ -268,17 +303,36 @@ resource "aws_lb_listener_rule" "unauthenticated_access_for_cidr_blocks" {
 
   condition {
     source_ip {
-      values = sort(compact(concat(var.allow_github_webhooks ? var.github_webhooks_cidr_blocks : [], var.whitelist_unauthenticated_cidr_blocks)))
+      values = local.whitelist_unauthenticated_cidr_block_chunks[count.index]
     }
   }
 }
 
-###################
+# Forward action for certain URL paths to bypass authentication (eg. GitHub webhooks)
+resource "aws_lb_listener_rule" "unauthenticated_access_for_webhook" {
+  count = var.allow_unauthenticated_access && var.allow_github_webhooks ? 1 : 0
+
+  listener_arn = module.alb.https_listener_arns[0]
+  priority     = var.allow_unauthenticated_webhook_access_priority
+
+  action {
+    type             = "forward"
+    target_group_arn = module.alb.target_group_arns[0]
+  }
+
+  condition {
+    path_pattern {
+      values = ["/events"]
+    }
+  }
+}
+
+################################################################################
 # Security groups
-###################
+################################################################################
 module "alb_https_sg" {
   source  = "terraform-aws-modules/security-group/aws//modules/https-443"
-  version = "v3.17.0"
+  version = "v4.3.0"
 
   name        = "${var.name}-alb-https"
   vpc_id      = local.vpc_id
@@ -292,7 +346,7 @@ module "alb_https_sg" {
 
 module "alb_http_sg" {
   source  = "terraform-aws-modules/security-group/aws//modules/http-80"
-  version = "v3.17.0"
+  version = "v4.3.0"
 
   name        = "${var.name}-alb-http"
   vpc_id      = local.vpc_id
@@ -306,7 +360,7 @@ module "alb_http_sg" {
 
 module "atlantis_sg" {
   source  = "terraform-aws-modules/security-group/aws"
-  version = "v3.17.0"
+  version = "v4.3.0"
 
   name        = var.name
   vpc_id      = local.vpc_id
@@ -318,7 +372,7 @@ module "atlantis_sg" {
       to_port                  = var.atlantis_port
       protocol                 = "tcp"
       description              = "Atlantis"
-      source_security_group_id = module.alb_https_sg.this_security_group_id
+      source_security_group_id = module.alb_https_sg.security_group_id
     },
   ]
 
@@ -327,12 +381,30 @@ module "atlantis_sg" {
   tags = merge(local.tags, var.atlantis_security_group_tags)
 }
 
-###################
+module "efs_sg" {
+  source  = "terraform-aws-modules/security-group/aws//modules/nfs"
+  version = "v4.8.0"
+  count   = var.enable_ephemeral_storage ? 0 : 1
+
+  name        = "${var.name}-efs"
+  vpc_id      = local.vpc_id
+  description = "Security group allowing access to the EFS storage"
+
+  ingress_cidr_blocks = [var.cidr]
+  ingress_with_source_security_group_id = [{
+    rule                     = "nfs-tcp",
+    source_security_group_id = module.atlantis_sg.security_group_id
+  }]
+
+  tags = local.tags
+}
+
+################################################################################
 # ACM (SSL certificate)
-###################
+################################################################################
 module "acm" {
   source  = "terraform-aws-modules/acm/aws"
-  version = "v2.12.0"
+  version = "v3.2.0"
 
   create_certificate = var.certificate_arn == ""
 
@@ -343,9 +415,9 @@ module "acm" {
   tags = local.tags
 }
 
-###################
-# Route53 record
-###################
+################################################################################
+# Route53 records
+################################################################################
 resource "aws_route53_record" "atlantis" {
   count = var.create_route53_record ? 1 : 0
 
@@ -354,27 +426,77 @@ resource "aws_route53_record" "atlantis" {
   type    = "A"
 
   alias {
-    name                   = module.alb.this_lb_dns_name
-    zone_id                = module.alb.this_lb_zone_id
+    name                   = module.alb.lb_dns_name
+    zone_id                = module.alb.lb_zone_id
     evaluate_target_health = true
   }
 }
 
-###################
+resource "aws_route53_record" "atlantis_aaaa" {
+  count = var.create_route53_aaaa_record ? 1 : 0
+
+  zone_id = data.aws_route53_zone.this[0].zone_id
+  name    = var.route53_record_name != null ? var.route53_record_name : var.name
+  type    = "AAAA"
+
+  alias {
+    name                   = module.alb.lb_dns_name
+    zone_id                = module.alb.lb_zone_id
+    evaluate_target_health = true
+  }
+}
+
+################################################################################
+# EFS
+################################################################################
+
+resource "aws_efs_file_system" "this" {
+  count = var.enable_ephemeral_storage ? 0 : 1
+
+  creation_token = var.name
+}
+
+resource "aws_efs_mount_target" "this" {
+  # we coalescelist in order to specify the resource keys when we create the subnets using the VPC or they're specified for us.  This works around the for_each value depends on attributes which can't be determined until apply error
+  for_each = {
+    for k, v in zipmap(coalescelist(var.private_subnets, var.private_subnet_ids), local.private_subnet_ids) : k => v
+    if var.enable_ephemeral_storage == false
+  }
+
+  file_system_id  = aws_efs_file_system.this[0].id
+  subnet_id       = each.value
+  security_groups = [module.efs_sg[0].security_group_id, module.atlantis_sg.security_group_id]
+}
+
+resource "aws_efs_access_point" "this" {
+  count = var.enable_ephemeral_storage ? 0 : 1
+
+  file_system_id = aws_efs_file_system.this[0].id
+  posix_user {
+    gid = local.gid
+    uid = local.uid
+  }
+}
+
+################################################################################
 # ECS
-###################
+################################################################################
 module "ecs" {
   source  = "terraform-aws-modules/ecs/aws"
-  version = "v2.5.0"
+  version = "v3.3.0"
+
+  create_ecs = var.create_ecs_cluster
 
   name               = var.name
   container_insights = var.ecs_container_insights
 
   capacity_providers = ["FARGATE", "FARGATE_SPOT"]
 
-  default_capacity_provider_strategy = {
-    capacity_provider = var.ecs_fargate_spot ? "FARGATE_SPOT" : "FARGATE"
-  }
+  default_capacity_provider_strategy = [
+    {
+      capacity_provider = var.ecs_fargate_spot ? "FARGATE_SPOT" : "FARGATE"
+    }
+  ]
 
   tags = local.tags
 }
@@ -412,10 +534,10 @@ resource "aws_iam_role" "ecs_task_execution" {
 }
 
 resource "aws_iam_role_policy_attachment" "ecs_task_execution" {
-  count = length(var.policies_arn)
+  for_each = toset(local.policies_arn)
 
   role       = aws_iam_role.ecs_task_execution.id
-  policy_arn = element(var.policies_arn, count.index)
+  policy_arn = each.value
 }
 
 # ref: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/specifying-sensitive-data.html
@@ -470,7 +592,7 @@ resource "aws_iam_role_policy" "ecs_task_access_secrets" {
 
 module "container_definition_github_gitlab" {
   source  = "cloudposse/ecs-container-definition/aws"
-  version = "v0.45.2"
+  version = "v0.58.1"
 
   container_name  = var.name
   container_image = local.atlantis_image
@@ -491,7 +613,7 @@ module "container_definition_github_gitlab" {
   container_depends_on     = var.container_depends_on
   essential                = var.essential
   readonly_root_filesystem = var.readonly_root_filesystem
-  mount_points             = var.mount_points
+  mount_points             = local.mount_points
   volumes_from             = var.volumes_from
 
   port_mappings = [
@@ -527,7 +649,7 @@ module "container_definition_github_gitlab" {
 
 module "container_definition_bitbucket" {
   source  = "cloudposse/ecs-container-definition/aws"
-  version = "v0.45.2"
+  version = "v0.58.1"
 
   container_name  = var.name
   container_image = local.atlantis_image
@@ -592,6 +714,31 @@ resource "aws_ecs_task_definition" "atlantis" {
 
   container_definitions = local.container_definitions
 
+  dynamic "ephemeral_storage" {
+    for_each = var.enable_ephemeral_storage ? [1] : []
+
+    content {
+      size_in_gib = var.ephemeral_storage_size
+    }
+  }
+
+  dynamic "volume" {
+    for_each = var.enable_ephemeral_storage ? [] : [1]
+
+    content {
+      name = "efs-storage"
+      efs_volume_configuration {
+        file_system_id          = aws_efs_file_system.this[0].id
+        transit_encryption      = "ENABLED"
+        transit_encryption_port = 2999
+        authorization_config {
+          access_point_id = aws_efs_access_point.this[0].id
+          iam             = "ENABLED"
+        }
+      }
+    }
+  }
+
   tags = local.tags
 }
 
@@ -605,7 +752,7 @@ data "aws_ecs_task_definition" "atlantis" {
 
 resource "aws_ecs_service" "atlantis" {
   name    = var.name
-  cluster = module.ecs.this_ecs_cluster_id
+  cluster = local.ecs_cluster_id
 
   task_definition                    = "${var.name}:${local.latest_task_definition_rev}"
   desired_count                      = var.ecs_service_desired_count
@@ -618,7 +765,7 @@ resource "aws_ecs_service" "atlantis" {
 
   network_configuration {
     subnets          = local.private_subnet_ids
-    security_groups  = [module.atlantis_sg.this_security_group_id]
+    security_groups  = [module.atlantis_sg.security_group_id]
     assign_public_ip = var.ecs_service_assign_public_ip
   }
 
@@ -628,8 +775,18 @@ resource "aws_ecs_service" "atlantis" {
     target_group_arn = element(module.alb.target_group_arns, 0)
   }
 
+  dynamic "load_balancer" {
+    for_each = var.extra_load_balancers
+    content {
+      container_name   = load_balancer.value["container_name"]
+      container_port   = load_balancer.value["container_port"]
+      target_group_arn = load_balancer.value["target_group_arn"]
+    }
+  }
+
   dynamic "capacity_provider_strategy" {
     for_each = var.ecs_fargate_spot ? [true] : []
+
     content {
       capacity_provider = "FARGATE_SPOT"
       weight            = 100
@@ -642,12 +799,13 @@ resource "aws_ecs_service" "atlantis" {
   tags = var.use_ecs_old_arn_format ? null : local.tags
 }
 
-###################
+################################################################################
 # Cloudwatch logs
-###################
+################################################################################
 resource "aws_cloudwatch_log_group" "atlantis" {
   name              = var.name
   retention_in_days = var.cloudwatch_log_retention_in_days
+  kms_key_id        = var.cloudwatch_logs_kms_key_id
 
   tags = local.tags
 }