+++ modules/github-repository-webhook/main.tf
@@ -1,12 +1,13 @@
 provider "github" {
-  token        = var.github_token
-  organization = var.github_organization
+  base_url = var.github_base_url
+  token    = var.github_token
+  owner    = var.github_owner
 }
 
 resource "github_repository_webhook" "this" {
-  count = var.create_github_repository_webhook ? length(var.atlantis_allowed_repo_names) : 0
+  count = var.create_github_repository_webhook ? length(var.atlantis_repo_allowlist) : 0
 
-  repository = var.atlantis_allowed_repo_names[count.index]
+  repository = var.atlantis_repo_allowlist[count.index]
 
   configuration {
     url          = var.webhook_url