+++ modules/github-repository-webhook/variables.tf
@@ -4,20 +4,26 @@ variable "create_github_repository_webhook" {
   default     = true
 }
 
+variable "github_base_url" {
+  description = "Github base URL to use when creating webhook (when using GitHub Enterprise)"
+  type        = string
+  default     = null
+}
+
 variable "github_token" {
   description = "Github token to use when creating webhook"
   type        = string
   default     = ""
 }
 
-variable "github_organization" {
-  description = "Github organization to use when creating webhook"
+variable "github_owner" {
+  description = "Github owner to use when creating webhook"
   type        = string
   default     = ""
 }
 
-variable "atlantis_allowed_repo_names" {
-  description = "List of names of repositories which belong to the organization specified in `github_organization`"
+variable "atlantis_repo_allowlist" {
+  description = "List of names of repositories which belong to the owner specified in `github_owner`"
   type        = list(string)
 }