+++ modules/gitlab-repository-webhook/variables.tf
@@ -16,8 +16,8 @@ variable "gitlab_token" {
   default     = ""
 }
 
-variable "atlantis_allowed_repo_names" {
-  description = "List of names of repositories which belong to the organization specified in `gitlab_organization`"
+variable "atlantis_repo_allowlist" {
+  description = "List of names of repositories which belong to the `gitlab_base_url` specified"
   type        = list(string)
 }